% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flipdownWidgets.R
\name{flipdownWidgets-shiny}
\alias{flipdownWidgets-shiny}
\alias{flipdownWidgetsOutput}
\alias{renderFlipdownWidgets}
\title{Shiny bindings for flipdownWidgets}
\usage{
flipdownWidgetsOutput(outputId, width = "100\%", height = "400px")

renderFlipdownWidgets(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a flipdownWidgets}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\value{
\code{flipdownWidgetsOutput} returns a flipdownWidgets output element that can be included in user interface.
And \code{renderFlipdownWidgets} returns a function which will generate data for render.
}
\description{
Output and render functions for using flipdownWidgets within Shiny
applications and interactive Rmd documents.
}
